requirements.txt
Flask == 3.0.0
Flask-cors == 4.0.0
spacy == 3.7.2
word2number-es == 1.0.1
Unidecode == 1.3.7
unicode == 2.9
pyspellchecker  == 0.7.2
transformers == 4.34.1
torch == 2.1.0
librosa == 0.10.1
gunicorn == 22.0.0
unicode == 2.9
spacy-loggers == 1.0.5
spacy-legacy == 3.0.12
text-unidecode ==1.3